requirements.txt
numpy
anthropic>=0.7.0
openai>=1.0.0
datasets
python-dotenv
matplotlib
pyyaml
tqdm
